feat(collection): refactor collection module
