+++ src/collection/collection.gql.ts
@@ -1,4 +1,5 @@
 import { gql } from 'apollo-boost';
+
 import { ITEMS_PER_PAGE } from '../my-workspace/constants';
 
 // TODO: Reduce to only what we need.
