+++ src/collection/components/index.ts
@@ -1,15 +1,17 @@
-import FragmentAdd from './FragmentAdd';
-import FragmentDetail from './FragmentDetail';
-import FragmentEdit from './FragmentEdit';
+import FragmentAdd from './fragment/FragmentAdd';
+import FragmentDetail from './fragment/FragmentDetail';
+import FragmentEdit from './fragment/FragmentEdit';
+import CollectionStillsModal from './modals/CollectionStillsModal';
 import CutFragmentModal from './modals/CutFragmentModal';
 import ReorderCollectionModal from './modals/ReorderCollectionModal';
 import ShareCollectionModal from './modals/ShareCollectionModal';
 
 export {
+	CollectionStillsModal,
+	CutFragmentModal,
 	ReorderCollectionModal,
 	ShareCollectionModal,
-	FragmentEdit,
 	FragmentAdd,
 	FragmentDetail,
-	CutFragmentModal,
+	FragmentEdit,
 };
