+++ src/collection/helpers/validation.ts
@@ -2,9 +2,9 @@ import { compact, isNil } from 'lodash-es';
 
 import { Avo } from '@viaa/avo2-types';
 
-import { MAX_SEARCH_DESCRIPTION_LENGTH } from '../../constants';
 import { stripHtml } from '../../shared/helpers/formatters/strip-html';
-import { toSeconds } from '../../shared/helpers/parsers/duration';
+
+import { MAX_SEARCH_DESCRIPTION_LENGTH } from '../collection.const';
 
 type ValidationRule<T> = {
 	error: string | ((object: T) => string);
