+++ src/collection/views/CollectionDetail/CollectionDetail.scss
