+++ src/collection/views/index.ts
@@ -1,0 +1,13 @@
+import CollectionDetail from './CollectionDetail/CollectionDetail';
+import CollectionEdit from './CollectionEdit/CollectionEdit';
+import CollectionEditContent from './CollectionEdit/CollectionEditContent';
+import CollectionEditMetaData from './CollectionEdit/CollectionEditMetaData';
+import CollectionOverview from './CollectionOverview/CollectionOverview';
+
+export {
+	CollectionEditContent,
+	CollectionEditMetaData,
+	CollectionDetail,
+	CollectionOverview,
+	CollectionEdit,
+};
